refactors ui elements
