+++ src/index.html
@@ -22,10 +22,10 @@
           <div class="header">
             <h1 class="title">JS-Matrix</h1>
             <span class="toolbar">
-              <button class="stop">
+              <button id="stop">
                 <svg width="15" height="15" id="Layer_1" viewBox="0 0 506.1 506.1" xmlns="http://www.w3.org/2000/svg"><path d="m489.609 0h-473.118c-9.108 0-16.491 7.383-16.491 16.491v473.118c0 9.107 7.383 16.491 16.491 16.491h473.119c9.107 0 16.49-7.383 16.49-16.491v-473.118c0-9.108-7.383-16.491-16.491-16.491z"></path></svg>
               </button>
-              <button class="play">
+              <button id="play">
                 <svg width="20" height="20" viewBox="0 0 494.942 494.942" xmlns="http://www.w3.org/2000/svg"><path d="m35.353 0 424.236 247.471-424.236 247.471z"></path></svg>
               </button>
             </span>
