+++ src/matrix/index.ts
@@ -1,0 +1,7 @@
+import renderMatrix from './renderers'
+import addEventHandlers from './events'
+
+export function render(notes: string[]) {
+  renderMatrix(notes)
+  addEventHandlers()
+}
